+++ NseApiDTO/NsePutDTO.cs
@@ -9,7 +9,7 @@ public class NsePutDTO
     public double ChangeinOpenInterest { get; set; }
     public double PchangeinOpenInterest { get; set; }
     public double TotalTradedVolume { get; set; }
-    public double LastPrice { get; set; }
+    public decimal LastPrice { get; set; }
     public double Bidprice { get; set; }
     public double AskPrice { get; set; }
     public double UnderlyingValue { get; set; }
